ci: skip internal component's demo rollup
