+++ components/helptext/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@aurodesignsystem/auro-helptext",
-  "version": "4.1.0",
+  "version": "1.0.0",
   "description": "auro-helptext HTML custom element",
   "type": "module",
   "main": "dist/index.js",
@@ -44,7 +44,7 @@
     "build": "run-s build:sass bundler types",
     "build:sass": "sass --no-source-map --load-path=../../node_modules \"./src/styles/:./src/styles/\"",
     "build:watch": "nodemon -e scss,js --watch src --ignore \"*.css\" --ignore \"*-css.js\" --exec \"npm run build\"",
-    "bundler": "rollup -c node:@auro-formkit/config/rollup",
+    "bundler": "rollup -c node:@auro-formkit/config/main.rollup",
     "clean": "rm -rf dist build",
     "postCss:component": "node ../../node_modules/@aurodesignsystem/auro-library/scripts/build/postCss.mjs",
     "sass:render": "sass-render --load-path=../../node_modules 'src/**/*.css' -t ../../node_modules/@aurodesignsystem/auro-library/scripts/build/staticStyles-template.js",
