+++ components/helptext/src/styles/color.scss
@@ -1,0 +1,5 @@
+/* stylelint-disable selector-class-pattern */
+
+.helptext-wrapper {
+  color: var(--ds-auro-helptext-color);
+}
