+++ components/helptext/src/styles/style.scss
@@ -1,4 +1,4 @@
-// Copyright (c) 2024 Alaska Airlines. All right reserved. Licensed under the Apache-2.0 license
+// Copyright (c) 2025 Alaska Airlines. All right reserved. Licensed under the Apache-2.0 license
 // See LICENSE in the project root for license information.
 
 // ---------------------------------------------------------------------
@@ -7,11 +7,16 @@
 
 @import "@aurodesignsystem/design-tokens/dist/tokens/SCSSVariables";
 
+
+:host([error]) {
+  --ds-auro-helptext-color: var(--ds-color-text-error-default, #{$ds-color-text-error-default});
+}
+
 .helptext-wrapper {
   display: none;
   margin-top: var(--ds-size-50, $ds-size-50);
   margin-bottom: 0;
-  color: var(--ds-auro-help-text-color);
+  color: var(--ds-auro-helptext-color);
   font-size: var(--ds-text-body-size-xs, $ds-text-body-size-xs);
   line-height: var(--ds-size-200, $ds-size-200);
 }
